crimson/osd: add osd to crush when it boots
this maps the corresponding behavior of classic OSD, we could rely on
the start script to do this job. but statfs reported by object storage
is more accurate than what we have using `df`.

in future, we could extract this feature out of crimson-osd by querying
crimson-osd for the size of underlying storage. and prepare the mon
command using a script. but let's have this function builtin in
crimson-osd at this moment. so it's compatible with classic OSD.

Signed-off-by: Kefu Chai <kchai@redhat.com>